feat(storage): distinguish overlap and non-overlap compaction
